common: Refactor client parser
Moves the parser to separate file. Simplifies adding
a new btp client support. Fixes duplicated parsing that
was happening in inherited constructors of bot clients.
